Add stock management to library database
A 'stock' column has been added to the book table in the library database to track how many copies of each book are available. Two new functions - 'add_stock' and 'remove_stock' - have been added to the codebase to facilitate the addition and removal of book stock using an ISBN identifier.